Wave 4: 704 more real names; a declaration-count oracle; adjudicate thread_win32
NAMING. 704 placeholder symbols renamed (253 marked static) at the source of
truth, config/symbols.json, and propagated into semantic_matches.json,
semantic_data_matches.json and parked.json. Names are recovered rather than
invented wherever an oracle covers the function: the HCEA PDB compilands give
Bungie's own names with full C signatures, and the January linker-map symbol
atlas covers translation units the PDB dropped. stalerefs.py reports 0 stale
source references after propagation.

Other house-rule edits, all verified byte-neutral: `real` replaces raw `float`;
prototypes moved into their objects' own headers (new hud_nav_points.h);
file-scope functions given `static`; hand-rolled bit constants replaced with
the cseries.h FLAG spellings; declarations merged with their assignments.

A DECLARATION-COUNT ORACLE. Four prototypes added to source/math/real_math.h
(matrix3x3_multiply, matrix3x3_rotation_to_quaternion, random_math_initialize,
random_math_dispose) cost units.obj a completed function, and the bisect is
worth recording because it refutes the obvious reading. units.c calls none of
the four. Removing either pair restores EXACT; keeping both breaks it; writing
the same four declarations on four lines instead of ten still breaks it;
moving them to the end of the header still breaks it; and four unrelated dummy
declarations at the same position break it identically. So the perturbation is
neither semantic, nor textual, nor positional: it is the COUNT of function
declarations visible in the translation unit, which shifts C2's
definition-position-driven allocation and flips a tie in
_unit_preprocess_node_orientations (lcsd 6, 585/585 instructions, 0 masked
blocks). Two extra declarations are tolerated, four are not.

The four prototypes are byte-neutral for every caller that needs them
(vehicles 56/5, real_math 115/2, matrix_math 36/1 measure identically with and
without), and none of the four implicit declarations can miscompile: all take
pointer or void arguments and no caller uses a return value. They are warning
hygiene that costs a completed function, so they are reverted. Any future
shared-header edit must be measured against the whole board, not just the
object it was written for.

THREAD_WIN32 ADJUDICATED, not fixed by a lie. 0eb0d09aa removed a false
`extern struct mutex_reference transport_address_string;` -- correctly, because
transport_address_string is a char[256] in another translation unit -- and the
function went residual. It is now proven unreachable rather than unsolved. The
loop bound &thread_globals.mutex_references[MAXIMUM_MUTEX_REFERENCES].in_use is
one past the end of the static thread_globals object (image 3262248, 1544
bytes), so it falls inside the next .bss symbol, transport_address_string
(image 3263792). csplit recovers the target from the linked image and must
attribute a destination to the symbol containing it, so it spells relocation 2
as _transport_address_string+36; VC7 forms the address from a C expression
rooted in thread_globals and necessarily spells it _thread_globals+1580. Both
resolve to image 3263828. Everything else is identical: size 64, three
relocations at addresses 5/27/40 all of type 6, equal normalized bytes, and all
three resolved image destinations equal. objdiff independently scores the
function 100.0%.

No C source makes VC7 emit csplit's spelling, so this records a limitation of
target recovery rather than a codegen gap. tools/parked_functions.py gains a
csplit-relocation-alias class whose admission requires exactly that proof:
equal size, equal normalized bytes, equal relocation count/addresses/types, and
independently resolved image destinations that agree.

New reusable tooling: scratch/verdicts.py (per-function board snapshot),
scratch/regress.py (rename-aware regression check) and scratch/fnsbs.py
(side-by-side target/ours disassembly for any function), replacing three
one-off scripts.

Gates: build exit 0 with 15 parked ties validated; fakescan reports no new
smells (the 9 pre-existing ones in main.c and units.c:10100 remain owner
items); stalerefs 0. Board 5165 exact, 75 pure-C residuals, 2 ftol-blocked,
1 parked asm, zero unintended regressions.

+++ source/scenario/wind.c
@@ -437,9 +437,9 @@ void scenario_get_water_current(
 static void wind_variance_initialize(
 	void)
 {
-	long control_point_index;
-	long sample_index;
-	long axis_index;
+	short control_point_index;
+	short sample_index;
+	short axis_index;
 
 	for (control_point_index = 0; control_point_index < 8; control_point_index++)
 	{
